% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staple_binimg.R
\name{staple_bin_img}
\alias{staple_bin_img}
\title{Run STAPLE on a set of nifti images}
\usage{
staple_bin_img(x, set_origin = FALSE, ...)
}
\arguments{
\item{x}{Character vector of filenames or list of arrays/images}

\item{set_origin}{Should the origin be set to the same if they images are
\code{niftiImage}s}

\item{...}{Additional arguments to \code{\link{staple_bin_mat}}}
}
\value{
A list similar to \code{\link{staple_bin_mat}}, but
has a resulting image
}
\description{
Run STAPLE on a set of nifti images
}
\examples{

n = 5
r = 1000
sens = c(0.8, 0.9, 0.8, 0.5, 0.8)
spec = c(0.9, 0.75, 0.99, 0.98, 0.92)
set.seed(20171120)
n_1 = 200
n_0 = r - n_1
truth = c(rep(0, n_0), rep(1, n_1))
pred_1 = rbinom(n = n, size = n_1, prob = sens)
pred_0 = rbinom(n = n, size = n_0, prob = spec)
pred_0 = sapply(pred_0, function(n) {
   sample(c(rep(0, n), rep(1, n_0 -n)))
})
pred_1 = sapply(pred_1, function(n) {
   sample(c(rep(1, n), rep(0, n_1 -n)))
})
pred = rbind(pred_0, pred_1)
true_sens = colMeans(pred[ truth == 1, ])
true_spec = colMeans(1-pred[ truth == 0, ])
x = lapply(seq(n), function(x) {
   array(pred[, x], dim = c(10,10, 10))
 })
staple_out = staple_bin_img(x, set_origin = FALSE)

}
