% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staple_example_data.R
\name{staple_example_data}
\alias{staple_example_data}
\title{STAPLE Example Data}
\usage{
staple_example_data()
}
\value{
Character vector of filenames
}
\description{
STAPLE Example Data
}
\examples{
staple_example_data()
}
