% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staple.R
\name{staple}
\alias{staple}
\alias{staple.default}
\alias{staple.list}
\alias{staple.character}
\alias{staple.array}
\title{Generic STAPLE Algorithm}
\usage{
staple(x, ..., set_origin = FALSE)

\method{staple}{default}(x, ..., set_origin = FALSE)

\method{staple}{list}(x, ..., set_origin = FALSE)

\method{staple}{character}(x, ..., set_origin = FALSE)

\method{staple}{array}(x, ..., set_origin = FALSE)
}
\arguments{
\item{x}{a nxr matrix where there are n raters and r elements rated,
a list of images, or a character vector.  Note, \code{\link{readNifti}}
is used for image filenames}

\item{...}{Options for STAPLE, see \code{\link{staple_bin_mat}} and
\code{\link{staple_multi_mat}}}

\item{set_origin}{Should the origin be set to the same if x is a
set of images, including \code{niftiImage}s.}
}
\description{
Tries to do the correct STAPLE algorithm (binary/multi-class) for the
type of input (array/matrix/list of images/filenames of images)
}
